this was instantiating for double, not int

+++ core/vnl/io/Templates/vnl_io_sym_matrix+int-.cxx
@@ -1,2 +1,2 @@
 #include <vnl/io/vnl_io_sym_matrix.txx>
-VNL_IO_SYM_MATRIX_INSTANTIATE(double);
+VNL_IO_SYM_MATRIX_INSTANTIATE(int);
